chore: Update .gitignore and package.json
This commit adds the `prod.env` file to the `.gitignore` and updates the `dev` and `start` scripts in the `package.json` file to include the `-p 4000` flag, specifying the port number for the development and production servers. These changes improve the project's configuration and ensure consistent behavior across different environments.